+++ R/bias.det.R
@@ -185,6 +185,9 @@ allele.info<-function(X,x.norm=NULL,method=c("TMM","TMMex","MedR","QN"),logratio
   pvals<-do.call(rbind,pvals)
   pvals<-cbind(X[,1:3],pvals)
   pvals<-na.omit(pvals)
+  ht<-sig.hets(pvals,plot = FALSE, verbose = verbose)
+  pvals<-data.frame(pvals,eH.pval=ht[,"pval"],eH.delta=ht[,"delta"])
+
   return(pvals)
 }